fix(models): Float→Numeric(12,2) for monetary fields in deal.py

+++ app/models/deal.py
@@ -2,7 +2,7 @@
 Сделка (Deal) — фиксирует факт договорённости между грузовладельцем и перевозчиком.
 Хранит полную историю: статусы, даты, документы.
 """
-from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, Enum, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, Enum, ForeignKey, Text, Numeric
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 import enum
@@ -41,11 +41,11 @@ class Deal(Base):
     status       = Column(Enum(DealStatus), default=DealStatus.confirmed, index=True)
 
     # Финансы (фиксируем на момент сделки)
-    agreed_price          = Column(Float, nullable=True)   # согласованная цена (в currency)
+    agreed_price          = Column(Numeric(12, 2), nullable=True)   # согласованная цена (в currency)
     currency              = Column(String(3), default="GEL")  # GEL / USD
-    exchange_rate_snapshot = Column(Float, nullable=True)  # GEL/USD зафиксирован при создании
-    final_price_gel       = Column(Float, nullable=True)   # итог в GEL (снапшот)
-    final_price_usd       = Column(Float, nullable=True)   # итог в USD (снапшот)
+    exchange_rate_snapshot = Column(Numeric(12, 6), nullable=True)  # GEL/USD зафиксирован при создании
+    final_price_gel       = Column(Numeric(12, 2), nullable=True)   # итог в GEL (снапшот)
+    final_price_usd       = Column(Numeric(12, 2), nullable=True)   # итог в USD (снапшот)
 
     # Даты
     created_at       = Column(DateTime(timezone=True), server_default=func.now())
